requirements.txt
# Core dependencies (all free/open-source)
cryptography>=41.0.0
numpy>=1.24.0
psutil>=5.9.0
pillow>=10.0.0
pycryptodome>=3.18.0

# Homomorphic Encryption
tenseal>=0.3.10

# Advanced Encryption
libnacl>=1.9.2

# Optional dependencies
scapy>=2.5.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
